Move BloodCompatibilityDetailResponse to admin package
